Refactor variable initialization
