+++ src/semantic_stmt.c
@@ -16,6 +16,7 @@
 #include "consteval.h"
 #include "semantic_expr.h"
 #include "semantic_init.h"
+#include "semantic_var.h"
 #include "semantic_global.h"
 #include "ir_global.h"
 #include "symtable.h"
@@ -125,136 +126,6 @@ static int check_typedef_stmt(stmt_t *stmt, symtable_t *vars)
     return 1;
 }
 
-/* Helpers for variable initialization */
-/*
- * Emit IR for a static array initialized with constant values.  The
- * array contents are provided up front and written directly to the
- * global data section.
- */
-static int init_static_array(ir_builder_t *ir, const char *name,
-                             const long long *vals, size_t count)
-{
-    return ir_build_glob_array(ir, name, vals, count, 1, 0);
-}
-
-/*
- * Store constant values into a dynamic array variable.  Each element is
- * written with either a volatile or normal store depending on the
- * declaration.
- */
-static void init_dynamic_array(ir_builder_t *ir, const char *name,
-                               const long long *vals, size_t count,
-                               int is_volatile)
-{
-    for (size_t i = 0; i < count; i++) {
-        ir_value_t idxv = ir_build_const(ir, (int)i);
-        ir_value_t valv = ir_build_const(ir, vals[i]);
-        if (is_volatile)
-            ir_build_store_idx_vol(ir, name, idxv, valv);
-        else
-            ir_build_store_idx(ir, name, idxv, valv);
-    }
-}
-
-/*
- * Store a constant initializer into a struct field.  The address of the
- * field is computed using the provided base pointer and offset and the
- * constant value is written through that pointer.
- */
-static void init_struct_member(ir_builder_t *ir, ir_value_t base,
-                               size_t off, long long val)
-{
-    ir_value_t offv = ir_build_const(ir, (int)off);
-    ir_value_t addr = ir_build_ptr_add(ir, base, offv, 1);
-    ir_value_t valv = ir_build_const(ir, val);
-    ir_build_store_ptr(ir, addr, valv);
-}
-/*
- * Expand an initializer list for an array variable.  The initializer
- * entries are evaluated, converted to constant values and stored either
- * at sequential indices or at explicit positions.  The resulting array
- * is then emitted using the appropriate helper.
- */
-static int handle_array_init(stmt_t *stmt, symbol_t *sym, symtable_t *vars, ir_builder_t *ir)
-{
-    long long *vals;
-    if (!expand_array_initializer(stmt->var_decl.init_list, stmt->var_decl.init_count,
-                                  sym->array_size, vars, stmt->line, stmt->column,
-                                  &vals))
-        return 0;
-    if (stmt->var_decl.is_static) {
-        if (!init_static_array(ir, sym->ir_name, vals, sym->array_size)) {
-            free(vals);
-            return 0;
-        }
-    }
-    else
-        init_dynamic_array(ir, sym->ir_name, vals, sym->array_size,
-                           stmt->var_decl.is_volatile);
-    free(vals);
-    return 1;
-}
-/*
- * Expand an initializer list for a struct variable.  Each provided
- * initializer assigns to the next field or to a named field.  Constant
- * values are stored into the computed member addresses of the struct
- * variable.
- */
-static int handle_struct_init(stmt_t *stmt, symbol_t *sym, symtable_t *vars, ir_builder_t *ir)
-{
-    long long *vals;
-    if (!expand_struct_initializer(stmt->var_decl.init_list, stmt->var_decl.init_count,
-                                   sym, vars, stmt->line, stmt->column, &vals))
-        return 0;
-    ir_value_t base = ir_build_addr(ir, sym->ir_name);
-    for (size_t i = 0; i < sym->struct_member_count; i++)
-        init_struct_member(ir, base, sym->struct_members[i].offset, vals[i]);
-    free(vals);
-    return 1;
-}
-
-/*
- * Compute layout information for an aggregate variable.  Struct and
- * union members are assigned offsets and the resulting element size is
- * stored back into the declaration.
- */
-static int compute_var_layout(stmt_t *stmt, symtable_t *vars)
-{
-    if (stmt->var_decl.type == TYPE_UNION) {
-        if (stmt->var_decl.member_count) {
-            size_t max = layout_union_members(stmt->var_decl.members,
-                                              stmt->var_decl.member_count);
-            stmt->var_decl.elem_size = max;
-        } else if (stmt->var_decl.tag) {
-            symbol_t *utype = symtable_lookup_union(vars, stmt->var_decl.tag);
-            if (!utype) {
-                error_set(stmt->line, stmt->column, error_current_file,
-                          error_current_function);
-                return 0;
-            }
-            stmt->var_decl.elem_size = utype->total_size;
-        }
-    }
-
-    if (stmt->var_decl.type == TYPE_STRUCT) {
-        if (stmt->var_decl.member_count) {
-            size_t total = layout_struct_members(
-                (struct_member_t *)stmt->var_decl.members,
-                stmt->var_decl.member_count);
-            stmt->var_decl.elem_size = total;
-        } else if (stmt->var_decl.tag) {
-            symbol_t *stype = symtable_lookup_struct(vars, stmt->var_decl.tag);
-            if (!stype) {
-                error_set(stmt->line, stmt->column, error_current_file,
-                          error_current_function);
-                return 0;
-            }
-            stmt->var_decl.elem_size = stype->struct_total_size;
-        }
-    }
-
-    return 1;
-}
 
 /*
  * Insert the variable into the given symbol table using the provided
@@ -346,234 +217,6 @@ static symbol_t *register_var_symbol(stmt_t *stmt, symtable_t *vars)
  * Allocates new member arrays and duplicates names. Returns non-zero on
  * success.
  */
-static int copy_union_metadata(symbol_t *sym, union_member_t *members,
-                               size_t count, size_t total)
-{
-    sym->total_size = total;
-    if (!count)
-        return 1;
-    sym->members = malloc(count * sizeof(*sym->members));
-    if (!sym->members)
-        return 0;
-    sym->member_count = count;
-    for (size_t i = 0; i < count; i++) {
-        union_member_t *m = &members[i];
-        sym->members[i].name = vc_strdup(m->name);
-        sym->members[i].type = m->type;
-        sym->members[i].elem_size = m->elem_size;
-        sym->members[i].offset = m->offset;
-        sym->members[i].bit_width = m->bit_width;
-        sym->members[i].bit_offset = m->bit_offset;
-    }
-    return 1;
-}
-
-/*
- * Copy struct member metadata from the parsed declaration to the symbol.
- * Allocates new member arrays and duplicates names. Returns non-zero on
- * success.
- */
-static int copy_struct_metadata(symbol_t *sym, struct_member_t *members,
-                                size_t count, size_t total)
-{
-    sym->struct_total_size = total;
-    if (!count)
-        return 1;
-    sym->struct_members = malloc(count * sizeof(*sym->struct_members));
-    if (!sym->struct_members)
-        return 0;
-    sym->struct_member_count = count;
-    for (size_t i = 0; i < count; i++) {
-        struct_member_t *m = &members[i];
-        sym->struct_members[i].name = vc_strdup(m->name);
-        sym->struct_members[i].type = m->type;
-        sym->struct_members[i].elem_size = m->elem_size;
-        sym->struct_members[i].offset = m->offset;
-        sym->struct_members[i].bit_width = m->bit_width;
-        sym->struct_members[i].bit_offset = m->bit_offset;
-        sym->struct_members[i].is_flexible = m->is_flexible;
-    }
-    return 1;
-}
-
-/*
- * Emit IR for a static initializer using a constant expression.
- * Handles scalars and aggregates placed in the global data section.
- */
-static int emit_static_initializer(stmt_t *stmt, symbol_t *sym,
-                                   symtable_t *vars, ir_builder_t *ir)
-{
-    long long cval;
-    if (!eval_const_expr(stmt->var_decl.init, vars, 0, &cval)) {
-        error_set(stmt->var_decl.init->line, stmt->var_decl.init->column, error_current_file, error_current_function);
-        return 0;
-    }
-    if (stmt->var_decl.type == TYPE_UNION)
-        ir_build_glob_union(ir, sym->ir_name, (int)sym->elem_size, 1,
-                           sym->alignment);
-    else if (stmt->var_decl.type == TYPE_STRUCT)
-        ir_build_glob_struct(ir, sym->ir_name, (int)sym->struct_total_size, 1,
-                            sym->alignment);
-    else
-        ir_build_glob_var(ir, sym->ir_name, cval, 1, sym->alignment);
-    return 1;
-}
-
-/*
- * Emit IR for a dynamic initializer evaluated at runtime.
- * The expression is checked for type compatibility and then stored.
- */
-static int emit_dynamic_initializer(stmt_t *stmt, symbol_t *sym,
-                                    symtable_t *vars, symtable_t *funcs,
-                                    ir_builder_t *ir)
-{
-    ir_value_t val;
-    type_kind_t vt = check_expr(stmt->var_decl.init, vars, funcs, ir, &val);
-    if (!(((is_intlike(stmt->var_decl.type) && is_intlike(vt)) ||
-           (is_floatlike(stmt->var_decl.type) &&
-            (is_floatlike(vt) || is_intlike(vt)))) ||
-          vt == stmt->var_decl.type)) {
-        error_set(stmt->var_decl.init->line, stmt->var_decl.init->column, error_current_file, error_current_function);
-        return 0;
-    }
-    if (stmt->var_decl.is_volatile)
-        ir_build_store_vol(ir, sym->ir_name, val);
-    else
-        ir_build_store(ir, sym->ir_name, val);
-    return 1;
-}
-
-/*
- * Emit IR for an initializer list of an aggregate variable.
- * Dispatches to the array or struct handler as appropriate.
- */
-static int emit_aggregate_initializer(stmt_t *stmt, symbol_t *sym,
-                                      symtable_t *vars, ir_builder_t *ir)
-{
-    if (stmt->var_decl.type == TYPE_ARRAY)
-        return handle_array_init(stmt, sym, vars, ir);
-    if (stmt->var_decl.type == TYPE_STRUCT)
-        return handle_struct_init(stmt, sym, vars, ir);
-    error_set(stmt->line, stmt->column, error_current_file, error_current_function);
-    return 0;
-}
-
-/*
- * Copy metadata describing aggregate members from the declaration to the
- * symbol table entry.  Dispatches to the struct or union helper based on the
- * variable's type.  Returns non-zero on success.
- */
-static int copy_aggregate_metadata(stmt_t *stmt, symbol_t *sym, symtable_t *vars)
-{
-    if (stmt->var_decl.type == TYPE_UNION)
-        return copy_union_metadata(sym, stmt->var_decl.members,
-                                   stmt->var_decl.member_count,
-                                   stmt->var_decl.elem_size);
-
-    if (stmt->var_decl.type == TYPE_STRUCT) {
-        if (stmt->var_decl.member_count == 0 && stmt->var_decl.tag) {
-            symbol_t *stype = symtable_lookup_struct(vars, stmt->var_decl.tag);
-            if (!stype)
-                return 0;
-            return copy_struct_metadata(sym, stype->struct_members,
-                                        stype->struct_member_count,
-                                        stype->struct_total_size);
-        }
-        return copy_struct_metadata(sym,
-                                    (struct_member_t *)stmt->var_decl.members,
-                                    stmt->var_decl.member_count,
-                                    stmt->var_decl.elem_size);
-    }
-
-    return 1;
-}
-
-/*
- * Evaluate the element count of a variable-length array and allocate
- * the required storage on the stack.  The resulting base pointer and
- * runtime length are stored in the symbol record.
- */
-static int handle_vla_size(stmt_t *stmt, symbol_t *sym, symtable_t *vars,
-                           symtable_t *funcs, ir_builder_t *ir)
-{
-    ir_value_t lenv;
-    if (check_expr(stmt->var_decl.size_expr, vars, funcs, ir, &lenv) ==
-        TYPE_UNKNOWN)
-        return 0;
-    ir_value_t eszv = ir_build_const(ir, (int)stmt->var_decl.elem_size);
-    ir_value_t total = ir_build_binop(ir, IR_MUL, lenv, eszv);
-    sym->vla_addr = ir_build_alloca(ir, total);
-    sym->vla_size = lenv;
-    return 1;
-}
-
-/*
- * Case 1: initialization with a constant expression.  The value can be
- * emitted directly into the data section and requires no runtime code.
- */
-static int emit_static_init(stmt_t *stmt, symbol_t *sym, symtable_t *vars,
-                            ir_builder_t *ir)
-{
-    return emit_static_initializer(stmt, sym, vars, ir);
-}
-
-/*
- * Case 2: scalar initializer evaluated at runtime for automatic storage.
- * The initializer expression is validated and stored through generated IR.
- */
-static int emit_dynamic_init(stmt_t *stmt, symbol_t *sym, symtable_t *vars,
-                             symtable_t *funcs, ir_builder_t *ir)
-{
-    return emit_dynamic_initializer(stmt, sym, vars, funcs, ir);
-}
-
-/*
- * Case 3: initializer list for an array or struct.  Elements are expanded
- * and stored sequentially via the aggregate initializer helper.
- */
-static int emit_aggregate_init(stmt_t *stmt, symbol_t *sym,
-                               symtable_t *vars, ir_builder_t *ir)
-{
-    return emit_aggregate_initializer(stmt, sym, vars, ir);
-}
-
-/*
- * Emit IR for any initializer attached to the variable.
- * Copies aggregate member metadata to the symbol and writes
- * constant or computed values using the IR builder.
- * Returns non-zero on success.
- */
-static int emit_var_initializer(stmt_t *stmt, symbol_t *sym,
-                                symtable_t *vars, symtable_t *funcs,
-                                ir_builder_t *ir)
-{
-    enum {
-        INIT_NONE,
-        INIT_STATIC,
-        INIT_DYNAMIC,
-        INIT_AGGREGATE
-    } kind = INIT_NONE;
-
-    if (!copy_aggregate_metadata(stmt, sym, vars))
-        return 0;
-
-    if (stmt->var_decl.init)
-        kind = stmt->var_decl.is_static ? INIT_STATIC : INIT_DYNAMIC;
-    else if (stmt->var_decl.init_list)
-        kind = INIT_AGGREGATE;
-
-    switch (kind) {
-    case INIT_STATIC:
-        return emit_static_init(stmt, sym, vars, ir);
-    case INIT_DYNAMIC:
-        return emit_dynamic_init(stmt, sym, vars, funcs, ir);
-    case INIT_AGGREGATE:
-        return emit_aggregate_init(stmt, sym, vars, ir);
-    default:
-        return 1;
-    }
-}
-
 static int check_var_decl_stmt(stmt_t *stmt, symtable_t *vars,
                                symtable_t *funcs, ir_builder_t *ir)
 {
